warning for "chat" pretrained; disable buggy evalita configs (#3127)
* check for chat for warning

* add test

* remove yaml extension from some evalita configs

* move unitxt to own test script

* fix CI test
